requirements.txt
-i https://pypi.org/simple
django-cors-headers==3.5.0
django==2.2
djangorestframework==3.11.1
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
